Minor fix over legend config
