+++ src/charts/areaChart/config.js
@@ -1,0 +1,16 @@
+/** 
+ * config.js
+ * @version:1.0.0
+ * @createdOn:12-Dec-2017
+ * @author:SmartChartsNXT
+ * @description:This will be the default Pie chart config for load in BaseChart.
+ */
+"use strict";
+
+export default function() {
+  let self = this; 
+  return {
+    minWidth: 250,
+    minHeight: 400
+  };
+}
