+++ src/charts/donutChart/config.js
@@ -10,7 +10,7 @@
 export default function() {
   let self = this; 
   return {
-    minWidth: (!self.props.opts.legends || self.props.opts.legends.enable !== false) && (self.props.opts.legends.float === 'left' || self.props.opts.legends.float === 'right') ? 500 : 300,
-    minHeight: (!self.props.opts.legends || self.props.opts.legends.enable !== false) && (self.props.opts.legends.float === 'top' || self.props.opts.legends.float === 'bottom') ? 500 : 400
+    minWidth: (!self.props.opts.legends ) || (self.props.opts.legends.enable !== false && (self.props.opts.legends.float === 'left' || self.props.opts.legends.float === 'right')) ? 500 : 300,
+    minHeight: (!self.props.opts.legends) || (self.props.opts.legends.enable !== false && (self.props.opts.legends.float === 'top' || self.props.opts.legends.float === 'bottom')) ? 500 : 400
   };
 }
